feat: refactored our record batch max + formatting
