+++ src/daft-distributed/src/pipeline_node/join/asof_join.rs
@@ -5,10 +5,11 @@ use common_metrics::{
     Meter,
     ops::{NodeCategory, NodeType},
 };
+use daft_core::{array::ops::build_multi_array_compare, datatypes::UInt64Array};
 use daft_dsl::expr::bound_expr::BoundExpr;
 use daft_local_plan::{LocalNodeContext, LocalPhysicalPlan, ShuffleReadBackend};
 use daft_logical_plan::{partitioning::HashClusteringConfig, stats::StatsState};
-use daft_recordbatch::{RecordBatch, record_batch_max_composite};
+use daft_recordbatch::RecordBatch;
 use daft_schema::schema::SchemaRef;
 use futures::{TryStreamExt, future::try_join_all};
 
@@ -452,6 +453,34 @@ impl PipelineNodeImpl for AsofJoinNode {
         ));
         TaskBuilderStream::from(result_rx)
     }
+}
+
+fn record_batch_max_composite(
+    batch: &RecordBatch,
+    composite_keys: &[BoundExpr],
+) -> DaftResult<Option<RecordBatch>> {
+    use std::cmp::Ordering;
+
+    if batch.is_empty() {
+        return Ok(None);
+    }
+
+    let cols: Vec<_> = composite_keys
+        .iter()
+        .map(|k| batch.eval_expression(k))
+        .collect::<DaftResult<_>>()?;
+    let descending = vec![false; composite_keys.len()];
+    let nulls_first = vec![true; composite_keys.len()];
+    let cmp = build_multi_array_compare(&cols, &descending, &nulls_first)?;
+    let max_idx = (1..batch.len()).fold(0usize, |best, i| {
+        if cmp(i, best) == Ordering::Greater {
+            i
+        } else {
+            best
+        }
+    });
+    let idx = UInt64Array::from_vec("idx", vec![max_idx as u64]);
+    Ok(Some(batch.take(&idx)?))
 }
 
 /// Creates top_n(limit=1, descending=true) tasks per (bucket, worker) pair to compute
