+++ src/daft-recordbatch/src/lib.rs
@@ -47,9 +47,7 @@ mod probeable;
 mod repr_html;
 
 pub use growable::GrowableRecordBatch;
-pub use ops::{
-    build_left_to_right_map, get_column_by_name, get_columns_by_name, record_batch_max_composite,
-};
+pub use ops::{build_left_to_right_map, get_column_by_name, get_columns_by_name};
 pub use probeable::{ProbeState, Probeable, ProbeableBuilder, make_probeable_builder};
 
 #[cfg(feature = "python")]
